+++ playground/error-tracking/next-ts-app/src/app/page.tsx
@@ -1,14 +1,16 @@
 'use client'
 
-function throwException() {
-    throw new Error('New Error')
-}
+import Link from 'next/link'
+import posthog from 'posthog-js'
 
 export default function Home() {
     return (
         <div>
             <main style={{ display: 'flex', justifyContent: 'center', alignItems: 'center', height: '100vh' }}>
-                <button onClick={() => throwException()}>Send exception</button>
+                <Link href="./error?messsage=Rendering%20Error">
+                    <button>Generate rendering error</button>
+                </Link>
+                <button onClick={() => posthog.captureException(new Error('Programming error'))}>Send error</button>
             </main>
         </div>
     )