+++ react/src/components/index.ts
@@ -1,1 +1,6 @@
 export * from './PostHogFeature'
+export {
+    PostHogErrorBoundary,
+    PostHogErrorBoundaryProps,
+    PostHogErrorBoundaryFallbackProps,
+} from './PostHogErrorBoundary'